requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
openai-whisper>=20231117
pyannote.audio>=3.1.0
pydub>=0.25.1
anthropic>=0.18.0
soundfile>=0.12.1

# Audio processing
numpy>=1.24.0
scipy>=1.10.0

# API
flask>=3.0.0
werkzeug>=3.0.0
flask-swagger-ui>=4.11.1
flask-cors>=4.0.0

# System monitoring
psutil>=5.9.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0